Refactored files to split into sync and async
